% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure1.R
\name{figure1}
\alias{figure1}
\title{figure1}
\usage{
figure1(df_prev = read_prev(), years = c(2023, 2024))
}
\arguments{
\item{df_prev}{The data object}

\item{years}{the years to filter}
}
\value{
A plot
}
\description{
Produce Figure 1 in the main chapter
}
