% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{prev_by_SPA}
\alias{prev_by_SPA}
\title{prev_by_SPA}
\usage{
prev_by_SPA(df_prev = read_prev(), inferCC = TRUE)
}
\arguments{
\item{df_prev}{the prevlance dataset}

\item{inferCC}{should there be inference of the clonal complex from spa type?}
}
\value{
An aggrigated dataset by SPA type
}
\description{
Aggregate by SPA type and only keep those that have a SPA type.
}
