% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_AMR}
\alias{read_AMR}
\title{read_AMR}
\usage{
read_AMR(
  path = "inst/extdata/MRSA AMR 2023-2024.xlsx",
  sheet = "QUERY_FOR_FULL_AMR_ISOL_DAT_000",
  EUvet = c(Cefoxitin = "C", Chloramphenicol = "C", Ciprofloxacin = "B", Clindamycin =
    "C", Erythromycin = "C", `Fusidic acid` = "D", Gentamicin = "C", Kanamycin = "C",
    Linezolid = "A", Mupirocin = "A", Penicillin = "D", `Quinupristin/Dalfopristin` =
    "A", Rifampicin = "A", Streptomycin = "C", Sulfamethoxazole = "D", Tetracycline =
    "D", Tiamulin = "C", Trimethoprim = "D", Vancomycin = "A"),
  AntibioticClass = c(Cefoxitin = "Cephalosporin", Chloramphenicol = "Amphenicol",
    Ciprofloxacin = "Fluoroquinolone", Clindamycin = "Lincosamide", Erythromycin =
    "Macrolide", `Fusidic acid` = "Steroid antibacterial", Gentamicin = "Aminoglycoside",
    Kanamycin = "Aminoglycoside", Linezolid = "Oxazolidinone", Mupirocin =
    "Pseudomonic acid", Penicillin = "Natural penicillin", `Quinupristin/Dalfopristin` =
    "Streptogramin", Rifampicin = "Rifamycin", Streptomycin = "Aminoglycoside",
    Sulfamethoxazole = "Sulfonamide", Tetracycline = "Tetracycline", 
     Tiamulin =
    "Pleuromutilin", Trimethoprim = "Dihydrofolate reductase inhibitor", Vancomycin =
    "Glycopeptide"),
  spa2CC = c(`11` = 398, `34` = 398, `108` = 398, `571` = 398, `588` = 398, `1255` = 398,
    `1451` = 398, `1456` = 398, `1580` = 398, `1793` = 398, `2011` = 398, `2330` = 398,
    `2346` = 398, `2576` = 398, `2922` = 398, `5452` = 398, `6228` = 398, `6575` = 398,
    `10485` = 398, `19248` = 398, `899` = "CC1/CC398", `127` = 1, `2` = 5, `242` = 5, `8`
    = 8, `9` = 8, `1419` = 1, `1430` = 1, `10204` = 1, `1422` = 692, `3512` = 2343),
  spa2ST = c(`1419` = 9, `1430` = 9, `10204` = 9),
  ST2CC = c(`9` = 1, `8325` = 1, `22` = 22),
  AB = c(Gentamicin = "GEN", Kanamycin = "KAN", Streptomycin = "STR", Chloramphenicol =
    "CHL", Rifampicin = "RIF", Ciprofloxacin = "CIP", Erythromycin = "ERY", Clindamycin =
    "CLI", `Quinupristin/Dalfopristin` = "Q/D", Linezolid = "LZD", Tiamulin = "TIA",
    Mupirocin = "MUP", `Fusidic acid` = "FUS", Sulfamethoxazole = "SMX", Trimethoprim =
    "TMP", Tetracycline = "TET", Vancomycin = "VAN", Cefoxitin = "CEF", Penicillin =
    "PEN")
)
}
\arguments{
\item{path}{The local file path to the excel file}

\item{sheet}{The name of the sheet to read}

\item{EUvet}{The classification of AB into levels of importance for public health}

\item{AntibioticClass}{The classification of AB into functional groups}

\item{spa2CC}{Rules for how to classify isolates with spatype but no CC}

\item{spa2ST}{Rules for how to classify isoaltes with spatype but no ST}

\item{ST2CC}{Rules for how to classify isolates with ST but no CC}

\item{AB}{shortforms of AB}
}
\value{
A data.table object
}
\description{
Read and cleanup the AMR isolate level data
}
\details{
These are the level of clinical importance of each substance
according to EUvet. This is not currently used but perhaps in
the future we can use it for something in the summary:

This can be used for definition of multidrug
resistance. Resistance to greater than 3 classes is
MDR. However, only the aminoglycosides are a class together so
basically resistance to any 3 indicates MDR. For information:
Steptomycin may not be the same groups as the other
aminogycosides since different action but is usually considered
so.

spa-types t1419, t1430 and t10204 were associated to ST9 (EFSA,
2009a ; Hasman et al., 2011 ; Köck et al., 2013), and classified
as CC1 (PubMLST1).

STs 9 and 8325 were classified as CC1 by PubMLST1.
ST22 was classified as CC22 by PubMLST1.

The data comes from the EFSA sharepoint:
"Dokument/General/SAS Files, Text Forms, National reports/
 3. 2024 EUSR on AMR datasets-calendar year 2025/SAS/
 Combined 2023 and 2024 data/2025.07.23/MRSA AMR 2023-2024.xlsx"

In subsequent corrections of the data, newer versions of the file
could be in later dates of the same 'All years' directory. In
subsequent reporting years, the higher up directories will change
too
}
