% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_prev}
\alias{read_prev}
\title{read_prev}
\usage{
read_prev(
  path = "inst/extdata/prevalence MRSA all data.xlsx",
  years = c("2023", "2024")
)
}
\arguments{
\item{path}{The local file path to the excel file}

\item{years}{The years to be included in the filtered data}
}
\value{
a data.table object
}
\description{
A function that reads the MRSA prevalence data reported by member
states. This is the so called "flat file" that is located here:
}
\details{
"Dokument/General/SAS Files, Text Forms, National reports/
 3. 2024 EUSR on AMR datasets-calendar year 2025/SAS/
 All years/2025.07.23/prevalence MRSA all data.xlsx"

In subsequent corrections of the data, newer versions of the file
could be in later dates of the same 'All years' directory. In
subsequent reporting years, the higher up directories will change
too
}
