% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2.R
\name{table2}
\alias{table2}
\title{table2}
\usage{
table2(df_AMR = read_AMR(), path_csv = tempfile(fileext = ".csv"))
}
\arguments{
\item{df_AMR}{The isolate based data object}

\item{path_csv}{path to the output csv file}

\item{years}{the years to filter}
}
\value{
A path to a csv file
}
\description{
Table 2 - Other types that are not CC398
Produce Table 2 in the main chapter
}
