% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_2.R
\name{table3}
\alias{table3}
\title{table3}
\usage{
table3(
  spaprev = prev_by_SPA(inferCC = FALSE),
  path_csv = tempfile(fileext = ".csv")
)
}
\arguments{
\item{spaprev}{The object from the prev_by_SPA() function}

\item{path_csv}{path to the output csv file}
}
\value{
A path to a csv file
}
\description{
Table 3 - Other types that are not CC classified Produce Table 3
that is not in the main chapter but helps to understand where the
missing data is.
}
