% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableE1.R
\name{tableE1}
\alias{tableE1}
\title{tableE1}
\usage{
tableE1(
  df_prev = read_prev(),
  year = 2024,
  path_csv = tempfile(fileext = ".csv")
)
}
\arguments{
\item{df_prev}{The data object}

\item{year}{the year to filter}

\item{path_csv}{path to a csv file}
}
\value{
path to a csv file
}
\description{
Produce Annex E Table 1: Methicillin-resistant Staphylococcus
aureus in food, 2024.
}
