% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableE3.R
\name{tableE3}
\alias{tableE3}
\title{tableE3}
\usage{
tableE3(
  df_prev = read_prev(),
  year = 2024,
  path_csv = tempfile(fileext = ".csv")
)
}
\arguments{
\item{df_prev}{The data object}

\item{year}{the year to filter}

\item{path_csv}{path to a csv file}
}
\value{
path to a csv file
}
\description{
Produce Annex E Table 3: Methicillin-resistant Staphylococcus
aureus in food-producing animals, clinical investigations
excluded, 2024.
}
