% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableE4.R
\name{tableE4}
\alias{tableE4}
\title{tableE3}
\usage{
tableE4(
  df_prev = read_prev(),
  year = 2023,
  path_csv = tempfile(fileext = ".csv")
)
}
\arguments{
\item{df_prev}{The data object}

\item{year}{the year to filter}

\item{path_csv}{path to a csv file}
}
\value{
path to a csv file
}
\description{
Produce Annex E table 4: Methicillin-resistant Staphylococcus
aureus in food-producing animals, clinical investigations
excluded, 2023.
}
