% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableE8_new.R
\name{tableE8_new}
\alias{tableE8_new}
\title{tableE8_new}
\usage{
tableE8_new(
  df_prev = read_prev(),
  year = 2024,
  inferCC = TRUE,
  path_csv = tempfile(fileext = ".csv")
)
}
\arguments{
\item{df_prev}{The isolate based data object}

\item{year}{the year to filter}

\item{inferCC}{should CCs be inferred from SPA}

\item{path_csv}{path to the output csv file}
}
\value{
A path to a csv file
}
\description{
Produce Annex Table 8: Methicillin-resistant Staphylococcus aureus
spa-type characterization, 2024.
}
