% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableE9.R
\name{tableE9}
\alias{tableE9}
\title{tableE9}
\usage{
tableE9(
  df_AMR = read_AMR(),
  year = 2023,
  path_csv = tempfile(fileext = ".csv")
)
}
\arguments{
\item{df_AMR}{The isolate based data object}

\item{year}{the year to filter}

\item{path_csv}{path to the output csv file}
}
\value{
A path to a csv file
}
\description{
Produce Annex Table 9: Methicillin-resistant Staphylococcus aureus
spa-type characterization, 2023.
}
