% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_2.R
\name{table2}
\alias{table2}
\title{table2}
\usage{
table2(
  spaprev = prev_by_SPA(inferCC = FALSE),
  path_csv = tempfile(fileext = ".csv")
)
}
\arguments{
\item{spaprev}{The object from the prev_by_SPA() function}

\item{path_csv}{path to the output csv file}
}
\value{
A path to a csv file
}
\description{
Table 2 - Other types that are not CC398
Produce Table 2 in the main chapter
}
